Move snackbar demo to /docs/snackbar
